Refactor BaseClient.Types, BaseClient docs
