+++ doc/data-modules.rst
@@ -1,8 +1,9 @@
 Data modules
 ============
 
-.. .. toctree::
-..    :maxdepth: 2
-..    :glob:
-..
-..    data-modules/*
+.. toctree::
+   :maxdepth: 2
+   :glob:
+
+   data-modules/*
+
