+++ test/unit/baseclient/types-suite.js
@@ -17,7 +17,7 @@ define(['require'], function(require) {
           set: function(path, value) {
             this._rootPaths[path] = value;
           }
-        }
+        };
 
       test.done();
     },
@@ -47,39 +47,6 @@ define(['require'], function(require) {
         }
       },
 
-      {
-        desc: "_attachType attaches default context to objects",
-        run: function(env, test) {
-          var obj = {
-            some: 'value'
-          };
-          env.storage = new RemoteStorage();
-          env.client = new BaseClient(env.storage, '/contacts/');
-          env.client._attachType(obj, 'contact');
-          test.assertAnd(obj, {
-            some: 'value',
-            '@context': 'http://remotestorage.io/spec/modules/contacts/contact'
-          });
-          test.done();
-        }
-      },
-
-      {
-        desc: "_attachType encodes special characters in type names for @context URI",
-        run: function(env, test) {
-          var obj = {
-            some: 'value'
-          };
-          env.storage = new RemoteStorage();
-          env.client = new BaseClient(env.storage, '/foo/');
-          env.client._attachType(obj, 'ba/F');
-          test.assertAnd(obj, {
-            some: 'value',
-            '@context': 'http://remotestorage.io/spec/modules/foo/ba%2FF'
-          });
-          test.done();
-        }
-      }
     ]
   });
